Remove view model of count resources use case
