--- centreon/src/Core/Resources/Infrastructure/API/CountResources/CountResourcesViewModel.php
@@ -1,74 +1,0 @@
-<?php
-
-/*
- * Copyright 2005 - 2025 Centreon (https://www.centreon.com/)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * For more information : contact@centreon.com
- *
- */
-
-declare(strict_types=1);
-
-namespace Core\Resources\Infrastructure\API\CountResources;
-
-/**
- * Class
- *
- * @class CountResourcesViewModel
- * @package Core\Resources\Infrastructure\API\CountResources
- */
-class CountResourcesViewModel {
-    /** @var int */
-    private int $totalFilteredResources = 0;
-
-    /** @var int */
-    private int $totalResources = 0;
-
-    /**
-     * @return int
-     */
-    public function getTotalFilteredResources(): int
-    {
-        return $this->totalFilteredResources;
-    }
-
-    /**
-     * @param int $totalFilteredResources
-     *
-     * @return void
-     */
-    public function setTotalFilteredResources(int $totalFilteredResources): void
-    {
-        $this->totalFilteredResources = $totalFilteredResources;
-    }
-
-    /**
-     * @return int
-     */
-    public function getTotalResources(): int
-    {
-        return $this->totalResources;
-    }
-
-    /**
-     * @param int $totalResources
-     *
-     * @return void
-     */
-    public function setTotalResources(int $totalResources): void
-    {
-        $this->totalResources = $totalResources;
-    }
-}
